Corrections:
Interval for segregation in [0.5,1] in metadata.json and in function checkparamseg in models.py

In models.py, in probmodel function the order of states in y1  must be the same of y1 in hoaxmodel to be compared correctly.  At each step of iteration y1 needs to be transformed in farray.

In odecomp.py the order of labels in the plot was wrong because of previous corrections.